backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
sqlalchemy==2.0.27
asyncpg==0.29.0
celery[redis]==5.3.6
openai==1.12.0
httpx==0.27.2
requests==2.31.0
google-generativeai
youtube-transcript-api==0.6.2
